added admin check for protected routes

+++ src/main/java/ch/zli/m223/punchclock/service/CategoryService.java
@@ -4,35 +4,39 @@ import ch.zli.m223.punchclock.domain.Category;
 import ch.zli.m223.punchclock.repository.CategoryRepository;
 import org.springframework.stereotype.Service;
 
+import java.security.Principal;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class CategoryService {
     private CategoryRepository categoryRepository;
+    private UserService userService;
 
-    public CategoryService(CategoryRepository categoryRepository) {
+    public CategoryService(CategoryRepository categoryRepository, UserService userService) {
         this.categoryRepository = categoryRepository;
+        this.userService = userService;
     }
 
     public List<Category> findAll() {
         return categoryRepository.findAll();
     }
 
-    public Optional<Category> findById(Long id) {
-        return categoryRepository.findById(id);
+    public Category create(Category category, Principal principal) throws Exception {
+        if (userService.isAdmin(principal)) {
+            return categoryRepository.saveAndFlush(category);
+        } else throw new Exception("User is not admin!");
     }
 
-    public Category create(Category Category) {
-        return categoryRepository.saveAndFlush(Category);
+    public void delete(Long id, Principal principal) throws Exception {
+        if (userService.isAdmin(principal)) {
+            categoryRepository.deleteById(id);
+        } else throw new Exception("User is not admin!");
     }
 
-    public void delete(Long id) {
-        categoryRepository.deleteById(id);
-    }
-
-    public Category update(Category Category) {
-        return categoryRepository.save(Category);
+    public Category update(Category category, Principal principal) throws Exception {
+        if (userService.isAdmin(principal)) {
+            return categoryRepository.save(category);
+        } else throw new Exception("User is not admin!");
     }
 
 }
